# docs/swagger.yaml
basePath: /api
definitions:
...
  cart.BasicCartOperationRequest:
    properties:
      cart_id:
        type: string
    type: object
...
  cart.CartMemberDTO:
    properties:
      joined_at:
        type: string
      user_id:
        type: integer
    type: object
  cart.CartResponse:
    properties:
      admin_id:
        type: integer
      cart_id:
        type: string
      items:
        items:
          $ref: '#/definitions/cart.CartItemDTO'
        type: array
      members:
        description: Только для shared
        items:
          $ref: '#/definitions/cart.CartMemberDTO'
        type: array
      mode:
        type: string
      restaurant_id:
        type: integer
      status:
        type: string
      total_cost:
        type: integer
    type: object
  cart.InviteResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
    type: object
...
  github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_cart.AddItemRequest:
    properties:
      cart_id:
        type: string
      dish_id:
        type: integer
      quantity:
        type: integer
    type: object
...
  github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_cart.RemoveItemRequest:
    properties:
      cart_id:
        type: string
      dish_id:
        type: integer
    type: object
  github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_cart.UpdateQuantityRequest:
    properties:
      cart_id:
        type: string
      dish_id:
        type: integer
      quantity:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API бэкенда для проекта "Delivery Club" от команды NaNcats
  title: NaNcats Delivery API
  version: "1.0"
paths:
...
  /api/cart:
    delete:
      consumes:
      - application/json
      description: Удаляет все товары из указанной корзины. В Shared-режиме доступно
        только админу.
      parameters:
      - description: Ключ идемпотентности
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/cart.BasicCartOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Корзина очищена
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет прав на выполнение операции
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Очистить корзину
      tags:
      - cart
    get:
      description: Возвращает текущую корзину авторизованного пользователя (включая
        статус комнаты и участников)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cart.CartResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить корзину
      tags:
      - cart
  /api/cart/close:
    post:
      consumes:
      - application/json
      description: Переводит корзину в Solo-режим. Удаляет всех участников кроме админа
        и удаляет их блюда.
      parameters:
      - description: Ключ идемпотентности
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: ID корзины
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/cart.BasicCartOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Корзина переведена в соло-режим
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Только админ может закрыть корзину
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Закрыть совместную корзину
      tags:
      - cart
  /api/cart/invite:
    post:
      description: Создает токен-приглашение в корзину. Переводит корзину в режим
        'shared'. Доступно только админу.
      parameters:
      - description: ID корзины
        in: query
        name: cart_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cart.InviteResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Только админ может генерировать ссылки
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сгенерировать инвайт-ссылку
      tags:
      - cart
  /api/cart/items:
    delete:
      consumes:
      - application/json
      description: Удаляет позицию из корзины. Гость может удалять только свое, админ
        - любое.
      parameters:
      - description: Ключ идемпотентности
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Данные для удаления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_cart.RemoveItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Товар удален
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет прав на удаление этой позиции
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить товар
      tags:
      - cart
    post:
      consumes:
      - application/json
      description: Добавляет товар в корзину. Вызывающий пользователь становится владельцем
        (owner) товара.
      parameters:
      - description: Ключ идемпотентности
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Данные товара
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_cart.AddItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Товар добавлен
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверное количество или блюдо из другого ресторана
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Пользователь не состоит в этой корзине
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить товар
      tags:
      - cart
    put:
      consumes:
      - application/json
      description: Изменяет количество существующего товара. Гость может менять только
        свои позиции, админ - любые.
      parameters:
      - description: Ключ идемпотентности
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Данные обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_cart.UpdateQuantityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Количество обновлено
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный запрос или количество
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет прав на изменение этой позиции
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить количество товара
      tags:
      - cart
  /api/cart/items/owner:
    patch:
      consumes:
      - application/json
      description: Переназначает блюдо на другого участника комнаты (или делает позицию
        ничейной, передав null). Только для админа.
      parameters:
      - description: Ключ идемпотентности
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Новый владелец
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_cart.ReassignOwnerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Владелец изменен
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Только админ может переназначать
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить плательщика товара
      tags:
      - cart
...
  /api/ws/cart:
    get:
      description: Устанавливает WS соединение для получения live-обновлений совместной
        корзины
      parameters:
      - description: ID корзины
        in: query
        name: cart_id
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
        "400":
          description: Bad Request (отсутствует cart_id)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: WebSocket соединение корзины
      tags:
      - cart
...
swagger: "2.0"